requirements.txt
git+https://github.com/huggingface/diffusers.git



transformers
accelerate
safetensors
sentencepiece
dashscope
kernels
torchvision
peft
torchao==0.11.0
pillow>=10.0.0
numpy>=1.24.0
diffusers
flask==3.0.0
runpod